moved the ServiceCompletionListener interface up to be reused by other Reader services
