+++ WordPress/src/main/java/org/wordpress/android/ui/reader/services/ServiceCompletionListener.java
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.reader.services.update;
+package org.wordpress.android.ui.reader.services;
 
 public interface ServiceCompletionListener {
     void onCompleted(Object companion);
